Refactoring of files and directories.
